divide the test into layers
