+++ src/render/__tests__/attributes.test.ts
@@ -1,23 +1,10 @@
-import { DotObject, GraphvizObject } from '../abstract';
-import { attribute } from '../../attribute';
-import { Attributes, AttributesBase } from '../attributes-base';
 import { toDot } from '../../render/to-dot';
+import { Attributes } from '../../model/attributes-base';
 
-describe('class Attributes', () => {
-  let attrs: Attributes<attribute.Attribute>;
+describe('Attributes rendering', () => {
+  let attrs: Attributes;
   beforeEach(() => {
     attrs = new Attributes();
-  });
-
-  it('should be instance of Attributes/AttributesBase/DotObject/GraphvizObject', () => {
-    expect(attrs).toBeInstanceOf(Attributes);
-    expect(attrs).toBeInstanceOf(AttributesBase);
-    expect(attrs).toBeInstanceOf(DotObject);
-    expect(attrs).toBeInstanceOf(GraphvizObject);
-  });
-
-  it('size should be 0 by default', () => {
-    expect(attrs.size).toBe(0);
   });
 
   describe('renders correctly by toDot method', () => {
@@ -43,25 +30,6 @@ describe('class Attributes', () => {
         fontsize: 16,
       });
       expect(toDot(attrs)).toMatchSnapshot();
-    });
-
-    test('apply/clear attribute', () => {
-      attrs.apply({
-        label: 'this is test',
-        color: 'red',
-        fontsize: 16,
-      });
-      expect(attrs.size).toBe(3);
-      attrs.clear();
-      expect(attrs.size).toBe(0);
-    });
-
-    test('set/get/delete attribute', () => {
-      const id = 'test';
-      attrs.set('label', id);
-      expect(attrs.get('label')).toBe(id);
-      attrs.delete('label');
-      expect(attrs.get('label')).toBeUndefined();
     });
 
     describe('edge with comment', () => {
